requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
